+++ packages/documentation/src/app/layout.tsx
@@ -1,6 +1,6 @@
-import {Footer, Layout, Link, Navbar} from "nextra-theme-docs"
 import {Banner, Head} from "nextra/components"
 import {getPageMap} from "nextra/page-map"
+import {Footer, Layout, Link, Navbar} from "nextra-theme-docs"
 import "nextra-theme-docs/style.css"
 import type {Metadata} from "next"
 import type {ReactNode} from "react"
