+++ packages/documentation/src/lib/playground.tsx
@@ -1,5 +1,17 @@
 "use client"
 
+import {zodResolver} from "@hookform/resolvers/zod"
+import {Editor, type Monaco, useMonaco} from "@monaco-editor/react"
+import {type Config, WebFsAdaptor} from "@nahkies/openapi-code-generator"
+import type {editor} from "monaco-editor"
+import type {AutoTypingsCore} from "monaco-editor-auto-typings/lib/AutoTypingsCore"
+import {Callout} from "nextra/components"
+import {useMDXComponents} from "nextra-theme-docs"
+import type React from "react"
+import type {PropsWithChildren} from "react"
+import {useCallback, useEffect, useRef, useState} from "react"
+import {useForm} from "react-hook-form"
+import {z} from "zod"
 import {LoadingSpinner} from "@/lib/loading-spinner"
 import {sampleFilenames} from "@/lib/playground/consts"
 import {ControlledSelect} from "@/lib/playground/controls/controlled-select"
@@ -11,20 +23,8 @@ import {
   subscribeToWorkerMessages,
 } from "@/lib/playground/worker/bridge"
 import type {Monad, WorkerResult} from "@/lib/playground/worker/types"
-import {zodResolver} from "@hookform/resolvers/zod"
-import {Editor, type Monaco, useMonaco} from "@monaco-editor/react"
-import {type Config, WebFsAdaptor} from "@nahkies/openapi-code-generator"
-import type {editor} from "monaco-editor"
-import type {AutoTypingsCore} from "monaco-editor-auto-typings/lib/AutoTypingsCore"
-import {useMDXComponents} from "nextra-theme-docs"
-import {Callout} from "nextra/components"
-import type React from "react"
-import type {PropsWithChildren} from "react"
-import {useCallback, useEffect, useRef, useState} from "react"
-import {useForm} from "react-hook-form"
-import {z} from "zod"
-import styles from "./playground.module.css"
 import {ConfigForm} from "./playground/config-form"
+import styles from "./playground.module.css"
 
 type IStandaloneCodeEditor = editor.IStandaloneCodeEditor
