+++ packages/documentation/src/lib/playground/controls/controlled-checkbox.tsx
@@ -1,4 +1,3 @@
-import React from "react"
 import {type UseControllerProps, useController} from "react-hook-form"
 
 export function ControlledCheckbox<
