+++ packages/documentation/src/lib/playground/controls/controlled-select.tsx
@@ -1,4 +1,4 @@
-import React, {useEffect} from "react"
+import {useEffect} from "react"
 import {type UseControllerProps, useController} from "react-hook-form"
 
 export function ControlledSelect<
