+++ packages/openapi-code-generator/src/typescript/common/schema-builders/zod-schema-builder.ts
@@ -3,7 +3,6 @@ import type {Reference} from "../../../core/openapi-types"
 import type {
   IRModel,
   IRModelArray,
-  IRModelBase,
   IRModelNumeric,
   IRModelString,
   MaybeIRModel,
