+++ packages/typescript-koa-runtime/src/joi.ts
@@ -10,7 +10,7 @@ export function parseRequestInput<Schema extends JoiSchema>(
   schema: Schema,
   input: unknown,
   type: RequestInputType,
-  // biome-ignore lint/suspicious/noExplicitAny: <explanation>
+  // biome-ignore lint/suspicious/noExplicitAny: needed
 ): any
 export function parseRequestInput(
   schema: undefined,
@@ -21,7 +21,7 @@ export function parseRequestInput<Schema extends JoiSchema>(
   schema: Schema | undefined,
   input: unknown,
   type: RequestInputType,
-  // biome-ignore lint/suspicious/noExplicitAny: <explanation>
+  // biome-ignore lint/suspicious/noExplicitAny: needed
 ): any {
   try {
     if (!schema) {
